Move Home components to separeted files
